fix(FluentLambdaClassInDifferentNamespace): move Address to separate file
