+++ src/M31.FluentApi.Tests/CodeGeneration/TestClasses/Abstract/FluentLambdaClassInDifferentNamespace/Student.cs
@@ -1,5 +1,4 @@
 // Non-nullable member is uninitialized
-
 #pragma warning disable CS8618
 
 using M31.FluentApi.Attributes;
@@ -15,21 +14,5 @@ namespace M31.FluentApi.Tests.CodeGeneration.TestClasses.Abstract.FluentLambdaCl
 
         [FluentLambda(1)]
         public Address Address { get; set; }
-    }
-}
-
-namespace SomeOtherNamespace
-{
-    [FluentApi]
-    public class Address
-    {
-        [FluentMember(0)]
-        public string HouseNumber { get; set; }
-
-        [FluentMember(1)]
-        public string Street { get; set; }
-
-        [FluentMember(2, "InCity")]
-        public string City { get; set; }
     }
 }
