+++ src/M31.FluentApi.Tests/CodeGeneration/TestDataProvider.cs
@@ -22,7 +22,7 @@ internal class TestDataProvider : IEnumerable<object[]>
                 new object[] { "Abstract", "EmptyClass", "Student" },
                 new object[] { "Abstract", "FluentDefaultMemberClass", "Student" },
                 new object[] { "Abstract", "FluentLambdaClass", "Student|Address" },
-                new object[] { "Abstract", "FluentLambdaClassInDifferentNamespace", "Student" },
+                new object[] { "Abstract", "FluentLambdaClassInDifferentNamespace", "Student|Address" },
                 new object[] { "Abstract", "FluentLambdaNullablePropertyClass", "Student" },
                 new object[] { "Abstract", "FluentLambdaRecursiveClass", "Student" },
                 new object[] { "Abstract", "FluentLambdaSingleStepClass", "Student" },
